moved out views and styles from models file
