+++ architecture_diagrams/orchestrator/specs.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Callable, Iterable, List, Optional, Protocol, Sequence, Set, Union, cast
 
-from architecture_diagrams.c4.model import ElementBase
+from architecture_diagrams.c4 import ElementBase
 from architecture_diagrams.c4.system_landscape import SystemLandscape
 from architecture_diagrams.extensions.relationships import RelationshipFilter
 
@@ -61,7 +61,7 @@ class ViewSpec:
 
     def build(self, model: SystemLandscape) -> None:
         # Create the view on the model and resolve includes/excludes
-        from architecture_diagrams.c4.model import ViewType
+        from architecture_diagrams.c4 import ViewType
 
         view = None
         if self.smart and self.view_type == ViewType.SYSTEM_LANDSCAPE:
@@ -126,7 +126,7 @@ class ViewSpec:
                     # For system landscape views, containers are not valid include targets.
                     # If a container was selected (e.g., "System/Container"), include its parent Software System instead.
                     try:
-                        from architecture_diagrams.c4.model import (
+                        from architecture_diagrams.c4 import (
                             ViewType as _VT,
                         )  # local import to avoid module-level cycles
                     except Exception:
