+++ notes.md
@@ -9,3 +9,4 @@ Introduce something like an importance parameter and filter on that
 Think about defining groups dynamically
 x Introduce management for different workspaces (to enable the above)
 Services connected with everything should not show edges and should be grouped into one (see above)
+Support manual layouting by exporting autogenerated one?
