+++ projects/banking_redis/views/inherited_views.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 
-from architecture_diagrams.c4.model import ViewType
+from architecture_diagrams.c4 import ViewType
 from architecture_diagrams.orchestrator.specs import IncludeRelByName, ViewSpec, derive_view
